requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
httpx>=0.27.0
mcp[cli]>=1.2.0
psycopg2-binary>=2.9.9
redis>=5.0.0
google-cloud-secret-manager>=2.20.0
PyJWT>=2.9.0
# Dev/test
pytest>=8.0.0
pytest-asyncio>=0.23.0
